dashboard: extract policy/storage/auto-update controls into shared components
Pull the tier-policy preset selector, L1 pruning card and automatic-updates
row out of their origin pages (filtering/basic, storage, system) into reusable
components so the same controls can be surfaced from Settings without forking
the mutation logic:

- components/filtering/PolicyProfileSelector (setPolicyProfile)
- components/settings/StoragePruningCard (setArchiveMode/setOperatorWindow/setPruneProfile)
- components/settings/AutoUpdateSection (setAutoUpdate)

Origin pages now import and render these; behaviour is unchanged.
