+++ dashboard/src/app/(dashboard)/system/page.tsx
@@ -18,8 +18,6 @@ import {
   useUpdateStatus,
   useStartUpdate,
   useRollbackUpdate,
-  useAutoUpdate,
-  useSetAutoUpdate,
   // Storage & Pruning
   useFullConfig,
   useNodeStatus,
@@ -37,6 +35,7 @@ import {
 } from "@/hooks/queries";
 import { getBackupDownloadUrl } from "@/lib/api/backup";
 import { useToast } from "@/components/ui/Toast";
+import { AutoUpdateSection } from "@/components/settings/AutoUpdateSection";
 import type { UpdateStatus } from "@/lib/api/system";
 import type { PruneProfile } from "@/types/api";
 import type { VerifyBackupResponse } from "@/types/api";
@@ -160,11 +159,6 @@ export default function SystemPage() {
   const startUpdate = useStartUpdate();
   const rollback = useRollbackUpdate();
 
-  // -- Auto-update opt-in (default OFF) --
-  const { data: autoUpdate } = useAutoUpdate();
-  const setAutoUpdateMut = useSetAutoUpdate();
-  const autoUpdateEnabled = autoUpdate?.enabled ?? false;
-
   const [confirmDialogOpen, setConfirmDialogOpen] = useState(false);
   const [rollbackDialogOpen, setRollbackDialogOpen] = useState(false);
 
@@ -261,23 +255,6 @@ export default function SystemPage() {
     } catch (err) {
       setIsUpdating(false);
       toastError("Update Failed", err instanceof Error ? err.message : "Unknown error");
-    }
-  };
-
-  const handleAutoUpdateToggle = async (next: boolean) => {
-    try {
-      const result = await setAutoUpdateMut.mutateAsync(next);
-      success(
-        result.enabled ? "Automatic Updates On" : "Automatic Updates Off",
-        result.enabled
-          ? "Signed releases will be verified and applied automatically (checked every 6h)."
-          : "This node will no longer update itself.",
-      );
-    } catch (err) {
-      toastError(
-        "Toggle Failed",
-        err instanceof Error ? err.message : "Could not change the auto-update setting",
-      );
     }
   };
 
@@ -584,40 +561,9 @@ export default function SystemPage() {
                 Check for Updates
               </Button>
 
-              {/* Automatic Updates (opt-in, default OFF) */}
+              {/* Automatic Updates (opt-in, default OFF). Shared with /settings/system. */}
               <div className="pt-4 border-t border-gray-800">
-                <div className="flex items-center justify-between">
-                  <div className="pr-4">
-                    <div className="flex items-center gap-2">
-                      <span className="text-sm font-medium text-gray-300">Automatic updates</span>
-                      <Badge variant={autoUpdateEnabled ? "success" : "default"}>
-                        {autoUpdateEnabled ? "On" : "Off"}
-                      </Badge>
-                    </div>
-                    <p className="text-xs text-gray-500 mt-0.5">
-                      When on, newer <strong>signed</strong> releases are applied automatically —
-                      the GPG signature and ghostd checksum are verified before any binary is
-                      swapped, and the node rolls back on a failed health check. Checked every 6h.
-                      Off by default.
-                    </p>
-                    {autoUpdate?.last_status?.result && (
-                      <p className="text-xs text-gray-500 mt-1">
-                        Last check: {autoUpdate.last_status.result}
-                        {autoUpdate.last_status.last_run
-                          ? ` · ${formatDate(autoUpdate.last_status.last_run)}`
-                          : ""}
-                        {autoUpdate.installed_version ? ` · installed ${autoUpdate.installed_version}` : ""}
-                        {autoUpdate.latest_version ? ` · latest ${autoUpdate.latest_version}` : ""}
-                      </p>
-                    )}
-                  </div>
-                  <Toggle
-                    enabled={autoUpdateEnabled}
-                    onChange={handleAutoUpdateToggle}
-                    disabled={setAutoUpdateMut.isPending}
-                    label="Toggle automatic updates"
-                  />
-                </div>
+                <AutoUpdateSection />
               </div>
 
               {/* Rollback */}
